refactor: create JavaFX project
