--- app/src/main/java/com/lask/app/App.java
@@ -1,13 +1,0 @@
-package com.lask.app;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
