--- app/src/main/java/com/lask/app/model/task/BasicTask.java
@@ -1,10 +1,0 @@
-package com.lask.app.model.task;
-
-public interface BasicTask extends Task {
-
-    /**
-     * setCompletionPercentage : set the percentage of completion of this task
-     */
-    void setCompletionPercentage(int percentage);
-
-}
