--- app/src/main/java/com/lask/app/model/task/BooleanTask.java
@@ -1,13 +1,0 @@
-package com.lask.app.model.task;
-
-/**
- * Represents a simple task with no progression state.
- */
-public interface BooleanTask extends Task {
-
-    /**
-     * setFinished : change the state of the task
-     * @param finished boolean, the new state
-     */
-    void setFinished(boolean finished);
-}
