+++ src/main/java/com/lask/App.java
@@ -1,0 +1,13 @@
+package com.lask;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
