+++ src/main/java/com/lask/model/task/BasicTask.java
@@ -1,0 +1,10 @@
+package com.lask.model.task;
+
+public interface BasicTask extends Task {
+
+    /**
+     * setCompletionPercentage : set the percentage of completion of this task
+     */
+    void setCompletionPercentage(int percentage);
+
+}
